remove publishmodal and location get from api
